Spit up the code into more logical places.
