+++ RexWindowProjcet/Assets/Editor/RexDiagnostics/Core/RexParser.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: edd8e864e83ad4846a31be01b4f1f50b
+timeCreated: 1459812145
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
