+++ RexWindowProjcet/RexWindowProjcet.CSharp.Editor.csproj
@@ -120,11 +120,12 @@
     <Compile Include="Assets\Editor\RexDiagnostics\Core\Helpers\MemberDetails.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\Helpers\MemberType.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\Helpers\NameSpaceInfo.cs" />
-    <Compile Include="Assets\Editor\RexDiagnostics\Core\Helpers\RexReflectionHelper.cs" />
+    <Compile Include="Assets\Editor\RexDiagnostics\Core\RexReflectionHelper.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\Helpers\UsingsHandler.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\Input\InputStateMachine.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\RexCompileEngine.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\RexHelper.cs" />
+    <Compile Include="Assets\Editor\RexDiagnostics\Core\RexParser.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Core\RexUtils.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Test\InputStateMachineTest.cs" />
     <Compile Include="Assets\Editor\RexDiagnostics\Test\IntelliSenceTests.cs" />
